feat(node): hook `malloc()` and `free()` to track native memory usage by this module and report that
v8 has a `v8::Isolate::AdjustAmountOfExternalAllocatedMemory` method to report native heap usage,
used to determine garbage collection frequency. because apr doesn't have any function to measure
memory usage, i need to hook into `malloc()` and `free()` to to this. note that c++ `new` and
`delete` will also call these functions so it also works with c++ part. but it will only track
functions that links to these hooked functions (that meas, only functions in this module), so other
modules are not affected, that's good.